Identify Spatially Homogeneous Regions with concordex

Compute the neighborhood consolidation matrix and identify spatial homogeneous regions.

Usage

calculateConcordex(x, ...)

# S4 method for class 'ANY'
calculateConcordex(
  x,
  labels,
  ...,
  n_neighbors = 30,
  compute_similarity = FALSE,
  BLUSPARAM,
  BNINDEX,
  BNPARAM = KmknnParam(),
  BPPARAM = SerialParam()
)

# S4 method for class 'SummarizedExperiment'
calculateConcordex(x, labels, ..., assay.type = "logcounts")

# S4 method for class 'SingleCellExperiment'
calculateConcordex(x, labels, ..., use.dimred = NULL)

# S4 method for class 'SpatialExperiment'
calculateConcordex(x, labels, ..., use.spatial = TRUE)

Arguments

x

A SpatialExperiment, SpatialFeatureExperiment, SingleCellExperiment, or SummarizedExperiment object containing a count matrix.

Otherwise, a numeric matrix-like object containing counts for observations (e.g. cells/spots) on the rows and features on the columns.

...

Other parameters passed to methods

labels

Observation labels used to compute the neighborhood consolidation matrix. Continuous or discrete labels are allowed, and typically, integer labels are assumed to be discrete.

Labels can be specified as follows:

A vector or matrix-like object with one entry per observation If a matrix, the observations should be on the rows and the label identifiers on the columns.

If x inherits from SummarizedExperiment-class, a string or character vector specifying the names of columns in colData(x) or the name of a dimensionality reduction result (see reducedDimNames)

n_neighbors

Number of neighbors to expect for each observation. Defaults to 30.

compute_similarity

Logical. Whether to return the label similarity matrix. Only useful if discrete labels are provided.

BLUSPARAM

A BlusterParam object specifying the clustering algorithm to use to identify spatial homogeneous regions. If this parameter is not specified, then regions are not returned. By default, this parameter is missing.

BNINDEX

A BiocNeighborIndex object containing the precomputed index information.

BPPARAM

A BiocParallelParam object specifying whether and how computing the metric for numerous observations shall be parallelized (see bpparam).

assay.type

String or integer scalar indicating the assay of x containing the counts.

use.dimred

Integer or string specifying the reduced dimensions to use for construction of the k-nearest neighbor graph. Note that if this is not NULL, reduced dimensions can not be used as labels to compute the neighborhood consolidation matrix.

use.spatial

Logical, should the spatial coordinates be used to compute the k-nearest neighbor graph?

Value

A sparse matrix